petit correctif sur le fonction UCB1 (il n'y avait pas de cast donc elle valait 1 tt le temps)

+++ Assets/Connect Four/scripts/Tree.cs
@@ -186,7 +186,7 @@ namespace ConnectFour
 			double maxValue = -1;
 			Node bestNode = null;
 			foreach (Node child in children.Keys) {
-				double evaluation = child.wins / child.plays + Math.Sqrt (2 * Math.Log (nbSimulation) / child.plays);
+        double evaluation = (double)child.wins / (double)child.plays + Math.Sqrt (2 * Math.Log ((double)nbSimulation) / (double)child.plays);
 				if (maxValue < evaluation) {
 					maxValue = evaluation;
 					bestNode = child;
